pyproject.toml
[project]
name = "oil-rag-dra"
version = "0.1.0"
description = "Bilingual RAG with Dynamic Retrieval Augmentation for Oil & Gas Reports"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "PyMuPDF>=1.23.0",
    "pyyaml>=6.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.100.0", "uvicorn>=0.23.0"]
web = ["streamlit>=1.25.0", "gradio>=3.40.0"]
dev = ["jupyter>=1.0.0", "pytest>=7.4.0", "black>=23.0.0", "flake8>=6.0.0"]
viz = ["matplotlib>=3.7.0", "seaborn>=0.12.0"]

[project.urls]
Homepage = "https://github.com/your-username/oil-rag-dra"
Documentation = "https://oil-rag-dra.readthedocs.io/"
Repository = "https://github.com/your-username/oil-rag-dra.git"
Issues = "https://github.com/your-username/oil-rag-dra/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["oil_rag*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]